+++ CHANGES.md
@@ -1,5 +1,6 @@
 #### [unreleased]
 * update Toolbar to ToolbarGroup and ToolbarButton
+* add support for block alignment toolbar
 
 #### 2.1.2 / 2020-09-15
 * use same _block name_ for oembed